+++ packages/eslint-plugin/src/rules/prefer-for-of.ts
@@ -58,7 +58,8 @@ export default util.createRule({
         node.type === AST_NODE_TYPES.BinaryExpression &&
         node.operator === '<' &&
         isMatchingIdentifier(node.left, name) &&
-        node.right.type === AST_NODE_TYPES.MemberExpression &&
+        (node.right.type === AST_NODE_TYPES.MemberExpression ||
+          node.right.type === AST_NODE_TYPES.OptionalMemberExpression) &&
         isMatchingIdentifier(node.right.property, 'length')
       ) {
         return node.right.object;
@@ -172,7 +173,8 @@ export default util.createRule({
         return (
           !contains(body, id) ||
           (node !== undefined &&
-            node.type === AST_NODE_TYPES.MemberExpression &&
+            (node.type === AST_NODE_TYPES.MemberExpression ||
+              node.type === AST_NODE_TYPES.OptionalMemberExpression) &&
             node.property === id &&
             sourceCode.getText(node.object) === arrayText &&
             !isAssignee(node))