+++ packages/eslint-plugin/src/util/misc.ts
@@ -11,14 +11,14 @@ import {
 /**
  * Check if the context file name is *.d.ts or *.d.tsx
  */
-export function isDefinitionFile(fileName: string): boolean {
+function isDefinitionFile(fileName: string): boolean {
   return /\.d\.tsx?$/i.test(fileName || '');
 }
 
 /**
  * Upper cases the first character or the string
  */
-export function upperCaseFirst(str: string): string {
+function upperCaseFirst(str: string): string {
   return str[0].toUpperCase() + str.slice(1);
 }
 
@@ -31,7 +31,7 @@ type InferOptionsTypeFromRuleNever<T> = T extends TSESLint.RuleModule<
 /**
  * Uses type inference to fetch the TOptions type from the given RuleModule
  */
-export type InferOptionsTypeFromRule<T> = T extends TSESLint.RuleModule<
+type InferOptionsTypeFromRule<T> = T extends TSESLint.RuleModule<
   string,
   infer TOptions
 >
@@ -41,7 +41,7 @@ export type InferOptionsTypeFromRule<T> = T extends TSESLint.RuleModule<
 /**
  * Uses type inference to fetch the TMessageIds type from the given RuleModule
  */
-export type InferMessageIdsTypeFromRule<T> = T extends TSESLint.RuleModule<
+type InferMessageIdsTypeFromRule<T> = T extends TSESLint.RuleModule<
   infer TMessageIds,
   unknown[]
 >
@@ -51,9 +51,7 @@ export type InferMessageIdsTypeFromRule<T> = T extends TSESLint.RuleModule<
 /**
  * Gets a string name representation of the given PropertyName node
  */
-export function getNameFromPropertyName(
-  propertyName: TSESTree.PropertyName,
-): string {
+function getNameFromPropertyName(propertyName: TSESTree.PropertyName): string {
   if (propertyName.type === AST_NODE_TYPES.Identifier) {
     return propertyName.name;
   }
@@ -61,9 +59,9 @@ export function getNameFromPropertyName(
 }
 
 /** Return true if both parameters are equal. */
-export type Equal<T> = (a: T, b: T) => boolean;
+type Equal<T> = (a: T, b: T) => boolean;
 
-export function arraysAreEqual<T>(
+function arraysAreEqual<T>(
   a: T[] | undefined,
   b: T[] | undefined,
   eq: (a: T, b: T) => boolean,
@@ -78,7 +76,7 @@ export function arraysAreEqual<T>(
 }
 
 /** Returns the first non-`undefined` result. */
-export function findFirstResult<T, U>(
+function findFirstResult<T, U>(
   inputs: T[],
   getResult: (t: T) => U | undefined,
 ): U | undefined {
@@ -95,8 +93,11 @@ export function findFirstResult<T, U>(
  * Gets a string name representation of the name of the given MethodDefinition
  * or ClassProperty node, with handling for computed property names.
  */
-export function getNameFromClassMember(
-  methodDefinition: TSESTree.MethodDefinition | TSESTree.ClassProperty,
+function getNameFromClassMember(
+  methodDefinition:
+    | TSESTree.MethodDefinition
+    | TSESTree.ClassProperty
+    | TSESTree.TSAbstractMethodDefinition,
   sourceCode: TSESLint.SourceCode,
 ): string {
   if (keyCanBeReadAsPropertyName(methodDefinition.key)) {
@@ -119,11 +120,25 @@ function keyCanBeReadAsPropertyName(
   );
 }
 
-export type ExcludeKeys<
+type ExcludeKeys<
   TObj extends Record<string, unknown>,
   TKeys extends keyof TObj
 > = { [k in Exclude<keyof TObj, TKeys>]: TObj[k] };
-export type RequireKeys<
+type RequireKeys<
   TObj extends Record<string, unknown>,
   TKeys extends keyof TObj
 > = ExcludeKeys<TObj, TKeys> & { [k in TKeys]-?: Exclude<TObj[k], undefined> };
+
+export {
+  arraysAreEqual,
+  Equal,
+  ExcludeKeys,
+  findFirstResult,
+  getNameFromClassMember,
+  getNameFromPropertyName,
+  InferMessageIdsTypeFromRule,
+  InferOptionsTypeFromRule,
+  isDefinitionFile,
+  RequireKeys,
+  upperCaseFirst,
+};