+++ packages/typescript-estree/src/node-utils.ts
@@ -22,7 +22,8 @@ const ASSIGNMENT_OPERATORS: ts.AssignmentOperator[] = [
 
 const LOGICAL_OPERATORS: (
   | ts.LogicalOperator
-  | ts.SyntaxKind.QuestionQuestionToken)[] = [
+  | ts.SyntaxKind.QuestionQuestionToken
+)[] = [
   SyntaxKind.BarBarToken,
   SyntaxKind.AmpersandAmpersandToken,
   SyntaxKind.QuestionQuestionToken,
@@ -610,41 +611,6 @@ export function convertTokens(ast: ts.SourceFile): TSESTree.Token[] {
   }
   walk(ast);
   return result;
-}
-
-/**
- * Get container token node between range
- * @param ast the AST object
- * @param start The index at which the comment starts.
- * @param end The index at which the comment ends.
- * @returns typescript container token
- * @private
- */
-export function getNodeContainer(
-  ast: ts.SourceFile,
-  start: number,
-  end: number,
-): ts.Node {
-  let container: ts.Node | null = null;
-
-  /**
-   * @param node the ts.Node
-   */
-  function walk(node: ts.Node): void {
-    const nodeStart = node.pos;
-    const nodeEnd = node.end;
-
-    if (start >= nodeStart && end <= nodeEnd) {
-      if (isToken(node)) {
-        container = node;
-      } else {
-        node.getChildren().forEach(walk);
-      }
-    }
-  }
-  walk(ast);
-
-  return container!;
 }
 
 export interface TSError {