+++ packages/typescript-estree/src/ts-estree/ts-estree.ts
@@ -472,6 +472,7 @@ interface ClassPropertyBase extends BaseNode {
   value: Expression | null;
   computed: boolean;
   static: boolean;
+  declare: boolean;
   readonly?: boolean;
   decorators?: Decorator[];
   accessibility?: Accessibility;
@@ -690,6 +691,7 @@ export interface ExportSpecifier extends BaseNode {
 export interface ExpressionStatement extends BaseNode {
   type: AST_NODE_TYPES.ExpressionStatement;
   expression: Expression;
+  directive?: string;
 }
 
 export interface ForInStatement extends BaseNode {
@@ -717,6 +719,7 @@ export interface ForStatement extends BaseNode {
 
 export interface FunctionDeclaration extends FunctionDeclarationBase {
   type: AST_NODE_TYPES.FunctionDeclaration;
+  body: BlockStatement;
 }
 
 export interface FunctionExpression extends FunctionDeclarationBase {
@@ -1060,7 +1063,6 @@ export interface TSConstructSignatureDeclaration extends FunctionSignatureBase {
 }
 
 export interface TSDeclareFunction extends FunctionDeclarationBase {
-  id: Identifier;
   type: AST_NODE_TYPES.TSDeclareFunction;
 }